Discontinue deCONZ Add-on (#122)
marthoc's deCONZ add-on is now discontinued. The official Hass.io add-on has surpassed this add-on in terms of functionality and will be more actively developed and updated to take advantage of new Hass.io features. Please use the official add-on instead. This add-on will not be updated past deCONZ 2.05.66 and any new features or PR's will be closed without comment.

Fixes #30. Fixes #95. Fixes #96. Fixes #103. Fixes #113. Fixes #114. Fixes #116. Fixes #118. Fixes #119. Fixes #121. Closes #76. Closes #80. Closes #112.